+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Update spells for patch 11.0.0
 -   Added cataclysm spells
 -   Fix cataclysm cooldowns
+-   Fix dispelable aura highlight
 
 ## [5.0.0] - 2024-07-27